Huge Review Refactor.

+++ model/request/review.go
@@ -4,8 +4,8 @@ type ReviewEntry struct {
 	Item string `json:"item,omitempty"`
 	Date string `json:"date,omitempty"`
 	Content string `json:"content,omitempty"`
-	Sentiment uint `json:"sentiment,omitempty"`
-	Stars uint `json:"stars,omitempty"`
+	Sentiment float64 `json:"sentiment,omitempty"`
+	Stars float64 `json:"stars,omitempty"`
 }
 
 type Review struct {
